% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_interface.R
\name{gdoc_checkout}
\alias{gdoc_checkout}
\title{Checkout a google document in Markdown format}
\usage{
gdoc_checkout(filename)
}
\arguments{
\item{filename}{a complete or partial filename to be searched for on your Google drive.}
}
\value{
a googledrive dribble augmented with a new column `local_md` and new class "markdown_doc"
}
\description{
Given a filename this funciton will search for a matching file on your Google drive.
If more than one match is found you will get a menu to choose from. If too many maches are
found the function will stop.
}
\details{
Once the target file has been identified it will be downloaded in docx format into a hidden
folder `.markdrive/` created under the current working directory. Pandoc is used to convert the
docx to .md and the result is placed in the current working directory. File metadata mapping
the .md to the remote source is saved in the markdrive folder.
}
\examples{
\dontrun{
 gdoc_checkout("GOT")
 #Check out a file matching "GOT", e.g. "my GOT theory"
}
}
