% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_interface.R
\name{gdoc_render}
\alias{gdoc_render}
\title{Create a Google Doc from a local md or Rmd}
\usage{
gdoc_render(filename, gdoc_name = NULL, gdoc_path = NULL)
}
\arguments{
\item{filename}{The path to the local file}

\item{gdoc_name}{A name to be given to the Google doc, if required.}

\item{gdoc_path}{A folder to place the Google doc in, if required.}
}
\value{
A handle to the created file.
}
\description{
Create a Google Doc from a local md or Rmd
}
