% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_interface.R, R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs

lhs \%>\% rhs
}
\description{
Pipe operator

See \code{magrittr::\link[magrittr]{\%>\%}} for details.
}
\keyword{internal}
